+++ ReleaseNotes.md
@@ -5,6 +5,8 @@
 ### Improvements
 * Updated all elements according to newest Figma version. 
 * SBBRadioButton and SBBCheckbox now reflect theming (color) changes.
+* Breaking change: Renaming of init parameters of SBBRadioButton and SBBCheckbox according to newest Figma version. 
+* Breaking change: SBBRadioButton and SBBCheckbox now feature an optional subtitle instead of customizable content.
 
 ## V0.5.2 - 12 November 2021